requirements.txt
asgiref==3.8.1
contourpy==1.1.1
cycler==0.12.1
Django==5.0.6
djangorestframework==3.15.2
fonttools==4.53.0
gunicorn==22.0.0
joblib==1.4.2
kiwisolver==1.4.5
matplotlib==3.9.0
numpy==1.26.4
packaging==24.1
pandas==2.2.2
pillow==10.3.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
scikit-learn==1.5.0
scipy==1.13.1
six==1.16.0
sqlparse==0.5.0
threadpoolctl==3.5.0
tzdata==2024.1
